Check for available IdPs in auth::is_configured
If there are no active IdPs, this should be regarded as non-configured
instead.

Fixes #551
