+++ login.php
@@ -35,6 +35,7 @@ require('setup.php');
 
 $wantsurl = optional_param('wantsurl', '', PARAM_LOCALURL); // Overrides $SESSION->wantsurl if given.
 if ($wantsurl !== '') {
+    // This is later used in core_login_get_return_url().
     $SESSION->wantsurl = (new moodle_url($wantsurl))->out(false);
 }
