+++ tests/locallib_test.php
@@ -470,94 +470,6 @@ class auth_saml2_locallib_testcase extends advanced_testcase {
         );
     }
 
-    /**
-     * Test test_is_configured
-     */
-    public function test_is_configured() {
-        global $DB;
-
-        $this->resetAfterTest();
-
-        // Add a fake IdP.
-        $url = 'http://www.example.com';
-        $DB->insert_record('auth_saml2_idps', array(
-            'metadataurl' => $url,
-            'entityid'    => $url,
-            'name'        => 'Test IdP',
-            'activeidp'   => 1));
-
-        /** @var auth_plugin_saml2 $auth */
-        $auth = get_auth_plugin('saml2');
-
-        $files = array(
-            'crt' => $auth->certcrt,
-            'pem' => $auth->certpem,
-            'xml' => $auth->get_file(md5($url) . '.idp.xml'),
-        );
-
-        // Setup, remove the phpuunit dataroot temp files for saml2.
-        foreach ($files as $file) {
-            if (file_exists($file)) {
-                @unlink($file);
-            }
-        }
-
-        $this->assertFalse($auth->is_configured());
-
-        // File crt: true.
-        // File pem: false.
-        // File xml: false.
-        // File result: failure.
-        touch($files['crt']);
-        $this->assertFalse($auth->is_configured());
-
-        // File crt: true.
-        // File pem: true.
-        // File xml: false.
-        // File result: failure.
-        touch($files['pem']);
-        $this->assertFalse($auth->is_configured());
-
-        // File crt: true.
-        // File pem: true.
-        // File xml: true.
-        // File result: success.
-        touch($files['xml']);
-        $this->assertTrue($auth->is_configured());
-    }
-
-    public function test_is_configured_works_with_multi_idp_in_one_xml() {
-        global $DB;
-
-        $this->resetAfterTest();
-
-        // Add two fake IdPs.
-        $metadataurl = 'https://idp.example.org/idp/shibboleth';
-        $DB->insert_record('auth_saml2_idps', array(
-            'metadataurl' => $metadataurl,
-            'entityid'    => 'https://idp1.example.org/idp/shibboleth',
-            'name'        => 'Test IdP 1',
-            'activeidp'   => 1));
-        $DB->insert_record('auth_saml2_idps', array(
-            'metadataurl' => $metadataurl,
-            'entityid'    => 'https://idp2.example.org/idp/shibboleth',
-            'name'        => 'Test IdP 2',
-            'activeidp'   => 1));
-
-        /** @var auth_plugin_saml2 $auth */
-        $auth = get_auth_plugin('saml2');
-
-        touch($auth->certcrt);
-        touch($auth->certpem);
-
-        $this->assertFalse($auth->is_configured());
-
-        $xmlfile = md5("https://idp.example.org/idp/shibboleth");
-        touch($auth->get_file("{$xmlfile}.idp.xml"));
-
-        $this->assertTrue($auth->is_configured());
-    }
-
     public function test_get_email_from_attributes() {
         $this->resetAfterTest();
